Finalized base ResourceManager implementation
- removed resource loaded race condition: if an entry is removed from the registry while we were checking it it's loaded we would access a deleted element from the hashtable
- removed unnecessary data structures from ResourceManger
- removed unnecessary locks from ResourceManager
- fixed how we search for the keyword in a resoure. before was looking for a substring in a cstring. changed to use the search algortihm function
- removed warnings in ResourceManager

+++ Memory/GeneralAllocator.h
@@ -16,7 +16,7 @@ static const int bytes_for_free_block = bytes_for_size + bytes_for_size + bytes_
 // ERROR STATES
 static constexpr bool GENERAL_ALLOCATOR_OUT_OF_MEMORY = false;
 
-typedef size_t Handle;
+typedef int64_t Handle;
 
 class GeneralAllocator {
 	typedef uint8_t byte;
